# gym-backend-api/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Weather API",
    "description": "JSON REST API for viewing weather data"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/create": {
      "post": {
        "summary": "Create a blog post",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Creates a user",
          "content": {
            "application/json": {
              "schema": {
                "post_title": "string",
                "post_content": "string",
                "post_user_id": "string"
              },
              "example": {
                "post_title": "Blog Post Title",
                "post_content": "Blog Post Content",
                "post_user_id": "1"
              }
            }
          }
        }
      }
    },
    "/blog": {
      "get": {
        "summary": "Get all blog posts",
        "description": "Endpoint to get all blog posts.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/blog/{id}": {
      "get": {
        "summary": "Get a blog post by its id",
        "description": "Endpoint to get a blog post by it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/blogs/{id}": {
      "get": {
        "summary": "Get all blog posts by user id",
        "description": "Endpoint to get all blog posts by user id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/update": {
      "patch": {
        "summary": "Update a blog post",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Updates a blog post by id",
          "content": {
            "application/json": {
              "schema": {
                "post_id": "string",
                "post_title": "string",
                "post_content": "string",
                "post_user_id": "string"
              },
              "example": {
                "post_id": "1",
                "post_title": "Blog Post Title",
                "post_content": "Blog Post Content",
                "post_user_id": "1"
              }
            }
          }
        }
      }
    },
    "/delete/{postID}": {
      "delete": {
        "summary": "Delete a blog post by id",
        "description": "Endpoint to delete a blog post by its id.",
        "parameters": [
          {
            "name": "postID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}